refactor: align with some common conventions
